refactor: package 생성
