--- src/main/java/racingcar/InputView.java
@@ -1,5 +1,0 @@
-package racingcar;
-
-public interface InputView {
-    void print(String view);
-}
