+++ src/main/java/racingcar/inputview/InputDevice.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.inputview;
 
 public interface InputDevice {
     int getIntNumber();
