+++ src/main/java/racingcar/inputview/InputView.java
@@ -1,0 +1,5 @@
+package racingcar.inputview;
+
+interface InputView {
+    void print(String view);
+}
