+++ src/test/java/racingcar/RacingCarTest.java
@@ -3,6 +3,7 @@ package racingcar;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.RacingCar;
 import racingcar.mock.MockAlwaysMaxValueRandomUtil;
 
 import static org.assertj.core.api.Assertions.assertThat;
